Graph Scattering Chapter is Written
This is a version of the graph scattering chapter that is close to
written.  While there are still many things that I want to change, most
of the material is in the chapter.  Basically, it won’t differ much
between this an the final version.  Also, it’s about 50 pages.